requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
requests>=2.26.0
llama-cpp-python>=0.1.0
psutil>=5.8.0
gputil>=1.4.0
